+++ pos-order/src/main/java/com/positivity/order/internal/client/InventoryResult.java
@@ -1,3 +1,5 @@
 package com.positivity.order.internal.client;
 
-public record InventoryResult(boolean sufficient, int availableQuantity) {}
+import java.math.BigDecimal;
+
+public record InventoryResult(boolean sufficient, BigDecimal availableQuantity) {}
